msm: mdss: fix UI scroll issue on 8909 target
When partial update is enabled, during adb shell
stop and start sometimes there is a mismatch between
width and height value of DSI controller and MDP, which
causes UI scroll like effect on screen. This change
fixes the MDP width and height initialization logic so that
for DSI and MDP, height and width parameters always match.

Change-Id: I85b90bfbe855e6744507df64a8622a1688462330
Signed-off-by: Sandeep Panda <spanda@codeaurora.org>