binder: privilege elevation vuln. in system_server
An elevation of privilege vulnerability in system_server could enable
local malicious application to execute arbitrary code within the
context of a privileged process. This issue is rated as High because
it could be used to gain local access to elevated capabilities, which
are not normally accessible to a third-party application.

CVE_2016_6674
Issue: PAELLA-297

Change-Id: Icc60d83aefdc55fb3502544cc00c7a4bbc0060b8